requirements.txt
pyjwt~=1.4.2
flask~=1.1.2
werkzeug~=1.0.1